apps/requirements.txt
Flask
Flask-bootstrap4
pylint
Flask-session
azure-cosmosdb-table
azure-cognitiveservices-vision-customvision
